bug fix: refactor

+++ seckill-account/src/main/java/com/jinshuai/seckill/account/dao/UserDao.java
@@ -19,6 +19,6 @@ public interface UserDao {
     List<User> listUsers();
 
     @Select("SELECT * FROM user WHERE id = #{id}")
-    User getUserById(int userId);
+    User getUserById(long userId);
 
 }
