+++ seckill-account/src/main/java/com/jinshuai/seckill/account/service/UserService.java
@@ -12,6 +12,6 @@ public interface UserService {
 
     List<User> listUsers();
 
-    User getUserById(int userId);
+    User getUserById(long userId);
 
 }
